atomic-steps/bugninja_v1_bad_approach/requirements.txt
patchright
openai
loguru
python-dotenv
pillow
pydantic
uuid 